Enhancement: allow ParaStation admin users to call mpiexec directly. Bugfix: fix blocking stdin in batch scripts (#189).
Update regression test suite:

commit 403b34288baf64677b9a015b2623668907bf11fe
Author: Dorian Krause <d.krause@fz-juelich.de>
Date:   Thu May 7 15:14:41 2015 +0200

    Add a test for ticket #189

+++ plugins/psslurm/src/psslurm/psslurmjob.c
@@ -251,7 +251,7 @@ static void clearTasks(struct list_head *taskList)
     }
 }
 
-PS_Tasks_t *findTaskByRank(struct list_head *taskList, uint16_t rank)
+PS_Tasks_t *findTaskByRank(struct list_head *taskList, int32_t rank)
 {
     list_t *pos, *tmp;
     PS_Tasks_t *task = NULL;
